경로 수정, 수정된 폴더들 반영
