+++ lib/screens/trade/components/more.dart
