+++ lib/screens/trade/trade_stuff_detail.dart
