+++ lib/screens/trade/trade_talent.dart
@@ -1,4 +1,4 @@
-import 'package:fling/trade/components/Trade_Talent_List.dart';
+import 'package:fling/screens/trade/components/Trade_Talent_List.dart';
 import 'package:flutter/material.dart';
 
 class TradeTal extends StatelessWidget {
