+++ lib/screens/trade/trade_talent_detail.dart
